judges/requirements.txt
aptos-sdk==0.10.0
fastapi
python-dotenv
pydantic
uvicorn